Revert "feat: APIにテスト環境を追加"
This reverts commit b4a042229209b0b2e0dd4a6c12d73261ef097705.
