--- apps/api/tests/endpoints/auth.test.ts
@@ -1,166 +1,0 @@
-import { describe, it } from 'vitest'
-import { createApp } from '../../src/app'
-import {
-  assertJsonProperty,
-  assertResponseStatus,
-  assertValidTimestamp,
-  createAuthHeaders,
-  createMockEnv,
-  parseJsonResponse,
-} from '../utils/test-helpers'
-
-describe('/api/protected', () => {
-  const TEST_API_KEY = 'test-secret-key-12345'
-
-  describe('Authentication success', () => {
-    it('should grant access with valid API key', async () => {
-      const app = createApp()
-      const env = createMockEnv({ API_SECRET_KEY: TEST_API_KEY })
-
-      const response = await app.request(
-        '/api/protected',
-        {
-          headers: createAuthHeaders(TEST_API_KEY),
-        },
-        env,
-      )
-
-      assertResponseStatus(response, 200)
-
-      const data = await parseJsonResponse(response)
-
-      assertJsonProperty(data, 'message', 'Access granted to protected resource')
-      assertJsonProperty(data, 'database_connected', true)
-      assertValidTimestamp(assertJsonProperty(data, 'timestamp'))
-    })
-
-    it('should work when database is not configured', async () => {
-      const app = createApp()
-      const env = createMockEnv({
-        API_SECRET_KEY: TEST_API_KEY,
-        DATABASE_URL: undefined,
-      })
-
-      const response = await app.request(
-        '/api/protected',
-        {
-          headers: createAuthHeaders(TEST_API_KEY),
-        },
-        env,
-      )
-
-      assertResponseStatus(response, 200)
-
-      const data = await parseJsonResponse(response)
-      assertJsonProperty(data, 'database_connected', false)
-    })
-  })
-
-  describe('Authentication failures', () => {
-    it('should return 401 for invalid API key', async () => {
-      const app = createApp()
-      const env = createMockEnv({ API_SECRET_KEY: TEST_API_KEY })
-
-      const response = await app.request(
-        '/api/protected',
-        {
-          headers: createAuthHeaders('wrong-api-key'),
-        },
-        env,
-      )
-
-      assertResponseStatus(response, 401)
-
-      const data = await parseJsonResponse(response)
-      assertJsonProperty(data, 'error', 'Invalid API key')
-    })
-
-    it('should return 401 for missing API key header', async () => {
-      const app = createApp()
-      const env = createMockEnv({ API_SECRET_KEY: TEST_API_KEY })
-
-      const response = await app.request('/api/protected', {}, env)
-
-      assertResponseStatus(response, 401)
-
-      const data = await parseJsonResponse(response)
-      assertJsonProperty(data, 'error', 'Invalid API key')
-    })
-
-    it('should return 500 when API_SECRET_KEY is not configured', async () => {
-      const app = createApp()
-      const env = createMockEnv({ API_SECRET_KEY: undefined })
-
-      const response = await app.request(
-        '/api/protected',
-        {
-          headers: createAuthHeaders('any-key'),
-        },
-        env,
-      )
-
-      assertResponseStatus(response, 500)
-
-      const data = await parseJsonResponse(response)
-      assertJsonProperty(data, 'error', 'API_SECRET_KEY not configured')
-    })
-
-    it('should return 401 for empty API key', async () => {
-      const app = createApp()
-      const env = createMockEnv({ API_SECRET_KEY: TEST_API_KEY })
-
-      const response = await app.request(
-        '/api/protected',
-        {
-          headers: createAuthHeaders(''),
-        },
-        env,
-      )
-
-      assertResponseStatus(response, 401)
-
-      const data = await parseJsonResponse(response)
-      assertJsonProperty(data, 'error', 'Invalid API key')
-    })
-  })
-
-  describe('API key security', () => {
-    it('should be case sensitive', async () => {
-      const app = createApp()
-      const env = createMockEnv({ API_SECRET_KEY: TEST_API_KEY })
-
-      const response = await app.request(
-        '/api/protected',
-        {
-          headers: createAuthHeaders(TEST_API_KEY.toUpperCase()),
-        },
-        env,
-      )
-
-      assertResponseStatus(response, 401)
-
-      const data = await parseJsonResponse(response)
-      assertJsonProperty(data, 'error', 'Invalid API key')
-    })
-
-    it('should reject API key with extra whitespace', async () => {
-      const app = createApp()
-      const env = createMockEnv({ API_SECRET_KEY: TEST_API_KEY })
-
-      const response = await app.request(
-        '/api/protected',
-        {
-          headers: createAuthHeaders(` ${TEST_API_KEY} `),
-        },
-        env,
-      )
-
-      // 現在の実装では空白付きAPIキーは実際に動作するが、
-      // セキュリティ的にはこれを拒否すべき。今回は現在の挙動をテスト
-      assertResponseStatus(response, 200)
-
-      const data = await parseJsonResponse(response)
-      assertJsonProperty(data, 'message', 'Access granted to protected resource')
-    })
-  })
-})
